+++ app/src/main/java/ru/okcode/icalc/command/operator/Multiplication.kt
@@ -1,6 +1,7 @@
 package ru.okcode.icalc.command.operator
 
 import ru.okcode.icalc.command.Operator
+import java.math.BigDecimal
 import javax.inject.Inject
 
 class Multiplication @Inject constructor() : Operator {
@@ -11,7 +12,7 @@ class Multiplication @Inject constructor() : Operator {
         return rang
     }
 
-    override fun calc(vararg args: Double): Double {
+    override fun calc(vararg args: BigDecimal): BigDecimal {
         val a = args[0]
         val b = args[1]
         return a.times(b)
